requirements.txt
pytest~=7.1.0
pytest-html
pytest-xdist 
pytest-ordering   
pytest-rerunfailures  
allure-pytest
PyYAML~=6.0